Replace find invocation, stylistic cleanup, exit
Replaces the `find` invocation with the pure-Perl File::Find. Also adjusted some of the existing style to be more in line with Perl::Critic and included a little additional error-handling as well as a graceful exit if no files are found. With any luck, I didn't even break anything, but it would be a good idea to double-check that the $key regex still works.